More unit tests, fix setLanguage
The setLanguage method had two problems which are now fixed:

(1) When an invalid language name was passed, a somewhat cryptic
AttributeError was raised. This is now wrapped in a more friendly
ValueError that says what you did wrong

(2) The QsciLexerXxxx instance was being created without a parent,
causing the reference to mysteriously disappear after the method exited
(and the self.lexer() method to later return None). Passing a parent
argument correctly allows this reference to hang around and be fetched
later. Yay for unit tests, for exposing this bug!