Overhaul enums
Wow, it turns out to be a lot more work than anticipated to get
enumerated types turned into plain strings. Converting from string to
enum is janky, but easy; converting the other way around is a real
hassle because each enum is just a glorified integer, and two unrelated
enums with the same integer value compare as equal.

So this commit uses a lookup based on the *type* of the enum, which
luckily is preserved (by sip, apparently), making it possible to
determine what kind of enum we're looking at even though it acts just
like a regular int. To make room for this added complexity, the help
strings were dropped--I'm not using them right now, and they introduce a
lot of clutter.

End result: `enum_value` now has a complement, `enum_string`.
Fixed the corresponding unit tests.